wallet: split wallet store itests to specific files
