--- wallet/internal/db/itest/wallet_store_test.go
@@ -1,1330 +1,0 @@
-//go:build itest
-
-package itest
-
-import (
-	"math"
-	"testing"
-	"time"
-
-	"github.com/btcsuite/btcwallet/wallet/internal/db"
-	"github.com/btcsuite/btcwallet/wallet/internal/db/page"
-	"github.com/stretchr/testify/require"
-)
-
-// uint32Ptr returns a pointer to the given uint32 value.
-func uint32Ptr(v uint32) *uint32 {
-	return &v
-}
-
-// newTestReq constructs a valid pagination request for tests.
-func newTestReq[C any](t *testing.T, limit uint32) page.Request[C] {
-	t.Helper()
-
-	req, err := page.NewRequest[C](limit)
-	require.NoError(t, err)
-
-	return req
-}
-
-// TestCreateWallet verifies that CreateWallet correctly creates a wallet
-// and returns its information.
-func TestCreateWallet(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-	params := CreateWalletParamsFixture("test-wallet")
-	info, err := store.CreateWallet(t.Context(), params)
-	require.NoError(t, err)
-	require.NotNil(t, info)
-
-	require.Equal(t, uint32(1), info.ID, "first wallet ID should be 1")
-	require.Equal(t, params.Name, info.Name)
-	require.Equal(t, params.IsImported, info.IsImported)
-	require.Equal(t, params.ManagerVersion, info.ManagerVersion)
-	require.Equal(t, params.IsWatchOnly, info.IsWatchOnly)
-
-	require.Nil(t, info.SyncedTo)
-	require.Nil(t, info.BirthdayBlock)
-	require.True(t, info.Birthday.IsZero())
-}
-
-// TestCreateWalletWithBirthday checks that CreateWallet correctly sets the
-// wallet's birthday timestamp.
-func TestCreateWalletWithBirthday(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	params := CreateWalletParamsFixture("birthday-wallet")
-	birthday := time.Now().UTC().Add(-30 * 24 * time.Hour)
-	params.Birthday = birthday
-
-	info, err := store.CreateWallet(t.Context(), params)
-	require.NoError(t, err)
-	require.NotNil(t, info)
-
-	require.Equal(t, birthday.Unix(), info.Birthday.Unix())
-	require.Nil(t, info.BirthdayBlock)
-}
-
-// TestCreateWalletDuplicateName verifies that creating a wallet with a
-// duplicate name fails with an appropriate error.
-func TestCreateWalletDuplicateName(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-	params := CreateWalletParamsFixture("duplicate-wallet")
-
-	_, err := store.CreateWallet(t.Context(), params)
-	require.NoError(t, err)
-
-	before := store.StatsSnapshot()
-
-	// Attempt to create second wallet with same name.
-	_, err = store.CreateWallet(t.Context(), params)
-	require.Error(t, err, "expected error creating duplicate wallet")
-	requireConstraintSQLError(t, err)
-
-	after := store.StatsSnapshot()
-	require.Equal(t, before.Unhealthy, after.Unhealthy)
-	require.Equal(t, before.RetryAttempts, after.RetryAttempts)
-	require.Equal(t, before.RetrySuccesses, after.RetrySuccesses)
-	require.Equal(t, before.RetryExhausted, after.RetryExhausted)
-	require.Equal(t, before.AmbiguousTxCommits, after.AmbiguousTxCommits)
-	require.Equal(t, before.Errors.Backend, after.Errors.Backend)
-	require.Equal(t, before.Errors.TotalErrs+1, after.Errors.TotalErrs)
-	require.Equal(
-		t,
-		before.Errors.PermanentErrs+1,
-		after.Errors.PermanentErrs,
-	)
-	require.Equal(t, before.Errors.Constraint+1, after.Errors.Constraint)
-	require.Equal(t, before.Errors.TransientErrs, after.Errors.TransientErrs)
-	require.Equal(t, before.Errors.FatalErrs, after.Errors.FatalErrs)
-}
-
-// TestCreateWalletVariants tests different wallet types.
-func TestCreateWalletVariants(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name   string
-		params func(string) db.CreateWalletParams
-	}{
-		{
-			name:   "imported wallet",
-			params: CreateImportedWalletParams,
-		},
-		{
-			name:   "watch-only wallet",
-			params: CreateWatchOnlyWalletParams,
-		},
-		{
-			name:   "standard wallet",
-			params: CreateWalletParamsFixture,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-
-			params := tc.params(tc.name)
-			store := NewTestStore(t)
-
-			info, err := store.CreateWallet(t.Context(), params)
-			require.NoError(t, err)
-			require.NotNil(t, info)
-			require.Equal(t, params.IsImported, info.IsImported)
-			require.Equal(t, params.IsWatchOnly, info.IsWatchOnly)
-		})
-	}
-}
-
-// TestGetWallet verifies that GetWallet retrieves a wallet by name.
-func TestGetWallet(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	params := CreateWalletParamsFixture("get-test-wallet")
-	created, err := store.CreateWallet(t.Context(), params)
-	require.NoError(t, err)
-
-	retrieved, err := store.GetWallet(t.Context(), params.Name)
-	require.NoError(t, err)
-	require.NotNil(t, retrieved)
-
-	require.Equal(t, created.ID, retrieved.ID)
-	require.Equal(t, created.Name, retrieved.Name)
-	require.Equal(t, created.IsImported, retrieved.IsImported)
-	require.Equal(t, created.ManagerVersion, retrieved.ManagerVersion)
-	require.Equal(t, created.IsWatchOnly, retrieved.IsWatchOnly)
-}
-
-// TestGetWalletNotFound verifies that GetWallet returns ErrWalletNotFound
-// when the wallet doesn't exist.
-func TestGetWalletNotFound(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-	before := store.StatsSnapshot()
-
-	_, err := store.GetWallet(t.Context(), "non-existent-wallet")
-	require.Error(t, err)
-	require.ErrorIs(t, err, db.ErrWalletNotFound)
-
-	after := store.StatsSnapshot()
-	require.Equal(t, before, after)
-}
-
-// TestListWallets verifies that ListWallets correctly returns wallets and
-// handles empty results without error.
-func TestListWallets(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	// Initially empty.
-	query := db.ListWalletsQuery{
-		Page: newTestReq[uint32](t, 10),
-	}
-
-	pageResult, err := store.ListWallets(t.Context(), query)
-	require.NoError(t, err)
-	require.Empty(t, pageResult.Items)
-	require.Nil(t, pageResult.Next)
-
-	// Create three wallets.
-	names := []string{"wallet-1", "wallet-2", "wallet-3"}
-	for _, name := range names {
-		params := CreateWalletParamsFixture(name)
-		_, err := store.CreateWallet(t.Context(), params)
-		require.NoError(t, err)
-	}
-
-	pageResult, err = store.ListWallets(t.Context(), query)
-	require.NoError(t, err)
-	require.Len(t, pageResult.Items, 3)
-
-	// Verify all names are present.
-	walletsName := make([]string, len(pageResult.Items))
-	for i, w := range pageResult.Items {
-		walletsName[i] = w.Name
-	}
-
-	require.ElementsMatch(t, names, walletsName)
-}
-
-// TestListWalletsZeroLimit verifies ListWallets rejects a zero page limit.
-func TestListWalletsZeroLimit(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-	before := store.StatsSnapshot()
-
-	_, err := store.ListWallets(t.Context(), db.ListWalletsQuery{})
-	require.ErrorIs(t, err, db.ErrInvalidPageLimit)
-
-	after := store.StatsSnapshot()
-	require.Equal(t, before, after)
-}
-
-// TestListWalletsPagination verifies that ListWallets paginates correctly and
-// sets Next without requiring an extra round-trip.
-func TestListWalletsPagination(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	names := []string{"wallet-1", "wallet-2", "wallet-3", "wallet-4"}
-	for _, name := range names {
-		params := CreateWalletParamsFixture(name)
-		_, err := store.CreateWallet(t.Context(), params)
-		require.NoError(t, err)
-	}
-
-	query := db.ListWalletsQuery{
-		Page: newTestReq[uint32](t, 2),
-	}
-
-	page1, err := store.ListWallets(t.Context(), query)
-	require.NoError(t, err)
-	require.Len(t, page1.Items, 2)
-	require.NotNil(t, page1.Next)
-	require.Equal(t, page1.Items[1].ID, *page1.Next)
-
-	query.Page.After = page1.Next
-	page2, err := store.ListWallets(t.Context(), query)
-	require.NoError(t, err)
-	require.Len(t, page2.Items, 2)
-	require.Nil(t, page2.Next)
-
-	query.Page.After = uint32Ptr(page2.Items[len(page2.Items)-1].ID)
-	page3, err := store.ListWallets(t.Context(), query)
-	require.NoError(t, err)
-	require.Empty(t, page3.Items)
-	require.Nil(t, page3.Next)
-
-	paged := append([]db.WalletInfo{}, page1.Items...)
-	paged = append(paged, page2.Items...)
-	require.Len(t, paged, len(names))
-
-	pagedNames := make([]string, len(paged))
-	for i, wallet := range paged {
-		pagedNames[i] = wallet.Name
-	}
-
-	require.Equal(t, names, pagedNames)
-}
-
-// TestListWalletsExactBoundary verifies that pagination correctly handles the
-// exact boundary case where total results equal page-size multiples.
-func TestListWalletsExactBoundary(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	names := []string{"wallet-1", "wallet-2", "wallet-3", "wallet-4"}
-	for _, name := range names {
-		_, err := store.CreateWallet(
-			t.Context(), CreateWalletParamsFixture(name),
-		)
-		require.NoError(t, err)
-	}
-
-	query := db.ListWalletsQuery{
-		Page: newTestReq[uint32](t, 2),
-	}
-
-	page1, err := store.ListWallets(t.Context(), query)
-	require.NoError(t, err)
-	require.Len(t, page1.Items, 2)
-	require.Equal(t, names[0], page1.Items[0].Name)
-	require.Equal(t, names[1], page1.Items[1].Name)
-	require.NotNil(t, page1.Next)
-	require.Equal(t, page1.Items[1].ID, *page1.Next)
-
-	query.Page.After = page1.Next
-	page2, err := store.ListWallets(t.Context(), query)
-	require.NoError(t, err)
-	require.Len(t, page2.Items, 2)
-	require.Equal(t, names[2], page2.Items[0].Name)
-	require.Equal(t, names[3], page2.Items[1].Name)
-	require.Nil(t, page2.Next)
-	require.Greater(t, page2.Items[0].ID, *page1.Next)
-
-	query.Page.After = uint32Ptr(page2.Items[len(page2.Items)-1].ID)
-	page3, err := store.ListWallets(t.Context(), query)
-	require.NoError(t, err)
-	require.Empty(t, page3.Items)
-	require.Nil(t, page3.Next)
-}
-
-// TestIterWallets verifies that IterWallets exhaustively retrieves all wallets
-// and yields the same results in the same order as manual cursor-based
-// pagination.
-func TestIterWallets(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	names := []string{"wallet-1", "wallet-2", "wallet-3", "wallet-4"}
-	for _, name := range names {
-		params := CreateWalletParamsFixture(name)
-		_, err := store.CreateWallet(t.Context(), params)
-		require.NoError(t, err)
-	}
-
-	query := db.ListWalletsQuery{
-		Page: newTestReq[uint32](t, 2),
-	}
-	expected := flattenWalletPages(collectWalletPages(t, store, query))
-
-	iterWallets := make([]db.WalletInfo, 0, len(expected))
-	for wallet, err := range store.IterWallets(t.Context(), query) {
-		require.NoError(t, err)
-
-		iterWallets = append(iterWallets, wallet)
-	}
-
-	require.Equal(t, expected, iterWallets)
-}
-
-// TestIterWalletsPaginated verifies that IterWallets produces the same
-// results as manual pagination and correctly signals end-of-list.
-func TestIterWalletsPaginated(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	names := []string{"wallet-1", "wallet-2", "wallet-3", "wallet-4"}
-	for _, name := range names {
-		params := CreateWalletParamsFixture(name)
-		_, err := store.CreateWallet(t.Context(), params)
-		require.NoError(t, err)
-	}
-
-	query := db.ListWalletsQuery{
-		Page: newTestReq[uint32](t, 2),
-	}
-
-	pages := collectWalletPages(t, store, query)
-	require.Len(t, pages, 2)
-	require.NotNil(t, pages[0].Next)
-	require.Nil(t, pages[1].Next)
-
-	expected := flattenWalletPages(pages)
-
-	iterWallets := make([]db.WalletInfo, 0, len(expected))
-	for wallet, err := range store.IterWallets(t.Context(), query) {
-		require.NoError(t, err)
-
-		iterWallets = append(iterWallets, wallet)
-	}
-
-	require.Equal(t, expected, iterWallets)
-}
-
-// TestListWalletsPagedFromCursor verifies that ListWallets can resume
-// pagination from a specific cursor position, returning only wallets after
-// that cursor.
-func TestListWalletsPagedFromCursor(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	names := []string{"wallet-1", "wallet-2", "wallet-3", "wallet-4"}
-
-	created := make([]*db.WalletInfo, 0, len(names))
-	for _, name := range names {
-		params := CreateWalletParamsFixture(name)
-		wallet, err := store.CreateWallet(t.Context(), params)
-		require.NoError(t, err)
-
-		created = append(created, wallet)
-	}
-
-	pageReq := newTestReq[uint32](t, 2)
-	pageReq.After = uint32Ptr(created[1].ID)
-
-	query := db.ListWalletsQuery{
-		Page: pageReq,
-	}
-
-	pageResult, err := store.ListWallets(t.Context(), query)
-	require.NoError(t, err)
-	require.Len(t, pageResult.Items, 2)
-	require.Equal(t, names[2], pageResult.Items[0].Name)
-	require.Equal(t, names[3], pageResult.Items[1].Name)
-	require.Nil(t, pageResult.Next)
-
-	query.Page.After = uint32Ptr(created[3].ID)
-	pageResult, err = store.ListWallets(t.Context(), query)
-	require.NoError(t, err)
-	require.Empty(t, pageResult.Items)
-	require.Nil(t, pageResult.Next)
-}
-
-// TestListWalletsPagedWithSyncMetadata verifies that paginated wallet
-// listings include sync metadata such as synced-to block and birthday block.
-func TestListWalletsPagedWithSyncMetadata(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-	queries := store.Queries()
-
-	birthday1 := time.Now().UTC().Add(-48 * time.Hour)
-	birthday2 := time.Now().UTC().Add(-24 * time.Hour)
-
-	params1 := CreateWalletParamsFixture("wallet-sync-1")
-	params1.Birthday = birthday1
-	wallet1, err := store.CreateWallet(t.Context(), params1)
-	require.NoError(t, err)
-
-	params2 := CreateWalletParamsFixture("wallet-sync-2")
-	params2.Birthday = birthday2
-	wallet2, err := store.CreateWallet(t.Context(), params2)
-	require.NoError(t, err)
-
-	block1 := CreateBlockFixture(t, queries, 100)
-	block2 := CreateBlockFixture(t, queries, 101)
-
-	err = store.UpdateWallet(t.Context(), db.UpdateWalletParams{
-		WalletID:      wallet1.ID,
-		SyncedTo:      &block2,
-		BirthdayBlock: &block1,
-	})
-	require.NoError(t, err)
-
-	err = store.UpdateWallet(t.Context(), db.UpdateWalletParams{
-		WalletID:      wallet2.ID,
-		SyncedTo:      &block2,
-		BirthdayBlock: &block1,
-	})
-	require.NoError(t, err)
-
-	query := db.ListWalletsQuery{
-		Page: newTestReq[uint32](t, 1),
-	}
-
-	page1, err := store.ListWallets(t.Context(), query)
-	require.NoError(t, err)
-	require.Len(t, page1.Items, 1)
-	require.NotNil(t, page1.Items[0].SyncedTo)
-	require.NotNil(t, page1.Items[0].BirthdayBlock)
-	require.False(t, page1.Items[0].Birthday.IsZero())
-	require.NotNil(t, page1.Next)
-
-	query.Page.After = page1.Next
-	page2, err := store.ListWallets(t.Context(), query)
-	require.NoError(t, err)
-	require.Len(t, page2.Items, 1)
-	require.NotNil(t, page2.Items[0].SyncedTo)
-	require.NotNil(t, page2.Items[0].BirthdayBlock)
-	require.False(t, page2.Items[0].Birthday.IsZero())
-	require.Nil(t, page2.Next)
-}
-
-// TestListWalletsDeterministicPagination verifies stable page ordering and
-// next-cursor behavior for multi-page wallet listings.
-func TestListWalletsDeterministicPagination(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-	names := []string{
-		"wallet-page-1",
-		"wallet-page-2",
-		"wallet-page-3",
-		"wallet-page-4",
-		"wallet-page-5",
-	}
-
-	for _, name := range names {
-		_, err := store.CreateWallet(
-			t.Context(), CreateWalletParamsFixture(name),
-		)
-		require.NoError(t, err)
-	}
-
-	pages := collectWalletPages(t, store, db.ListWalletsQuery{
-		Page: newTestReq[uint32](t, 2),
-	})
-	require.Len(t, pages, 3)
-	require.Len(t, pages[0].Items, 2)
-	require.Len(t, pages[1].Items, 2)
-	require.Len(t, pages[2].Items, 1)
-	require.NotNil(t, pages[0].Next)
-	require.NotNil(t, pages[1].Next)
-	require.Nil(t, pages[2].Next)
-
-	wallets := flattenWalletPages(pages)
-	require.Len(t, wallets, len(names))
-
-	collectedNames := make([]string, len(wallets))
-	for i, wallet := range wallets {
-		if i > 0 {
-			require.Less(t, wallets[i-1].ID, wallet.ID)
-		}
-
-		collectedNames[i] = wallet.Name
-	}
-
-	require.Equal(t, names, collectedNames)
-	require.Equal(t, wallets[1].ID, *pages[0].Next)
-	require.Equal(t, wallets[3].ID, *pages[1].Next)
-
-	seenIDs := make(map[uint32]struct{}, len(wallets))
-	for i := range pages {
-		for j, wallet := range pages[i].Items {
-			_, duplicate := seenIDs[wallet.ID]
-			require.False(t, duplicate)
-
-			seenIDs[wallet.ID] = struct{}{}
-
-			// Skip the first item on the first page; there's no prior cursor
-			// to compare against.
-			if i == 0 && j == 0 {
-				continue
-			}
-
-			// First item on a later page: verify it sorts strictly after the
-			// previous page's cursor to ensure no gaps or duplicates at page
-			// boundaries.
-			if j == 0 {
-				require.Greater(t, wallet.ID, *pages[i-1].Next)
-				continue
-			}
-
-			// Items within the same page: verify strict ordering to ensure
-			// the page contents are sorted.
-			require.Greater(t, wallet.ID, pages[i].Items[j-1].ID)
-		}
-	}
-}
-
-// TestListWalletsInsertAfterCursor verifies inserts after page N are returned
-// on page N+1 when pagination uses increasing ID cursors.
-func TestListWalletsInsertAfterCursor(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-	for _, name := range []string{"wallet-a", "wallet-b", "wallet-c"} {
-		_, err := store.CreateWallet(
-			t.Context(), CreateWalletParamsFixture(name),
-		)
-		require.NoError(t, err)
-	}
-
-	query := db.ListWalletsQuery{
-		Page: newTestReq[uint32](t, 2),
-	}
-	page1, err := store.ListWallets(t.Context(), query)
-	require.NoError(t, err)
-	require.Len(t, page1.Items, 2)
-	require.NotNil(t, page1.Next)
-
-	// Pagination is by increasing wallet ID (id > cursor).
-	// A wallet created after page 1 should therefore appear on page 2.
-	inserted, err := store.CreateWallet(
-		t.Context(), CreateWalletParamsFixture("wallet-d"),
-	)
-	require.NoError(t, err)
-
-	query.Page.After = page1.Next
-	page2, err := store.ListWallets(t.Context(), query)
-	require.NoError(t, err)
-	require.Len(t, page2.Items, 2)
-	require.Equal(t, "wallet-c", page2.Items[0].Name)
-	require.Equal(t, inserted.ID, page2.Items[1].ID)
-	require.Equal(t, "wallet-d", page2.Items[1].Name)
-	require.Nil(t, page2.Next)
-}
-
-// TestListWalletsCursorEdges verifies stale and zero-value cursors produce
-// deterministic page results.
-func TestListWalletsCursorEdges(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	names := []string{"wallet-1", "wallet-2", "wallet-3"}
-	for _, name := range names {
-		_, err := store.CreateWallet(
-			t.Context(), CreateWalletParamsFixture(name),
-		)
-		require.NoError(t, err)
-	}
-
-	staleReq := newTestReq[uint32](t, 2)
-	staleReq.After = uint32Ptr(math.MaxUint32)
-
-	stalePage, err := store.ListWallets(t.Context(), db.ListWalletsQuery{
-		Page: staleReq,
-	})
-	require.NoError(t, err)
-	require.Empty(t, stalePage.Items)
-	require.Nil(t, stalePage.Next)
-
-	zeroReq := newTestReq[uint32](t, 2)
-	zeroReq.After = uint32Ptr(0)
-
-	zeroPage, err := store.ListWallets(t.Context(), db.ListWalletsQuery{
-		Page: zeroReq,
-	})
-	require.NoError(t, err)
-	require.Len(t, zeroPage.Items, 2)
-	require.Equal(t, names[0], zeroPage.Items[0].Name)
-	require.Equal(t, names[1], zeroPage.Items[1].Name)
-	require.NotNil(t, zeroPage.Next)
-}
-
-// collectWalletPages collects paginated wallet results by iterating through all
-// pages from ListWallets until Next is nil.
-func collectWalletPages(t *testing.T, store db.WalletStore,
-	query db.ListWalletsQuery) []page.Result[db.WalletInfo, uint32] {
-
-	t.Helper()
-
-	pages := make([]page.Result[db.WalletInfo, uint32], 0)
-	for {
-		pageResult, err := store.ListWallets(t.Context(), query)
-		require.NoError(t, err)
-
-		pages = append(pages, pageResult)
-
-		if pageResult.Next == nil {
-			return pages
-		}
-
-		query.Page.After = pageResult.Next
-	}
-}
-
-// flattenWalletPages flattens paginated wallet results into a single
-// slice containing all wallets from all pages.
-func flattenWalletPages(
-	pages []page.Result[db.WalletInfo, uint32]) []db.WalletInfo {
-
-	count := 0
-	for i := range pages {
-		count += len(pages[i].Items)
-	}
-
-	wallets := make([]db.WalletInfo, 0, count)
-	for i := range pages {
-		wallets = append(wallets, pages[i].Items...)
-	}
-
-	return wallets
-}
-
-// TestUpdateWalletSyncedTo checks that updating the wallet's synced-to block
-// works correctly.
-func TestUpdateWalletSyncedTo(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-	queries := store.Queries()
-
-	params := CreateWalletParamsFixture("update-sync-wallet")
-	created, err := store.CreateWallet(t.Context(), params)
-	require.NoError(t, err)
-
-	block := CreateBlockFixture(t, queries, 100)
-
-	updateParams := db.UpdateWalletParams{
-		WalletID: created.ID,
-		SyncedTo: &block,
-	}
-	err = store.UpdateWallet(t.Context(), updateParams)
-	require.NoError(t, err)
-
-	retrieved, err := store.GetWallet(t.Context(), created.Name)
-	require.NoError(t, err)
-	require.NotNil(t, retrieved.SyncedTo)
-	require.Equal(t, block.Height, retrieved.SyncedTo.Height)
-
-	// Assert fields that were not updated remain unchanged.
-	require.Equal(t, created.ID, retrieved.ID)
-	require.Equal(t, created.Name, retrieved.Name)
-	require.Equal(t, created.IsImported, retrieved.IsImported)
-	require.Equal(t, created.ManagerVersion, retrieved.ManagerVersion)
-	require.Equal(t, created.IsWatchOnly, retrieved.IsWatchOnly)
-	require.Nil(t, retrieved.BirthdayBlock)
-}
-
-// TestUpdateWalletBirthdayBlock checks that updating the wallet's birthday
-// block works correctly.
-func TestUpdateWalletBirthdayBlock(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-	queries := store.Queries()
-
-	params := CreateWalletParamsFixture("update-birthday-wallet")
-	created, err := store.CreateWallet(t.Context(), params)
-	require.NoError(t, err)
-
-	// Initially, BirthdayBlock should be nil.
-	require.Nil(t, created.BirthdayBlock)
-
-	block := CreateBlockFixture(t, queries, 50)
-
-	updateParams := db.UpdateWalletParams{
-		WalletID:      created.ID,
-		BirthdayBlock: &block,
-	}
-	err = store.UpdateWallet(t.Context(), updateParams)
-	require.NoError(t, err)
-
-	retrieved, err := store.GetWallet(t.Context(), created.Name)
-	require.NoError(t, err)
-	require.NotNil(t, retrieved.BirthdayBlock)
-	require.Equal(t, block.Height, retrieved.BirthdayBlock.Height)
-	require.Equal(t, block.Hash, retrieved.BirthdayBlock.Hash)
-	require.Equal(t, block.Timestamp.Unix(),
-		retrieved.BirthdayBlock.Timestamp.Unix())
-
-	// Assert fields that were not updated remain unchanged.
-	require.Equal(t, created.ID, retrieved.ID)
-	require.Equal(t, created.Name, retrieved.Name)
-	require.Equal(t, created.IsImported, retrieved.IsImported)
-	require.Equal(t, created.ManagerVersion, retrieved.ManagerVersion)
-	require.Equal(t, created.IsWatchOnly, retrieved.IsWatchOnly)
-	require.Nil(t, retrieved.SyncedTo)
-}
-
-// TestUpdateWalletBirthday checks that updating the wallet's birthday
-// timestamp works correctly.
-func TestUpdateWalletBirthday(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	params := CreateWalletParamsFixture("birthday-timestamp-wallet")
-	created, err := store.CreateWallet(t.Context(), params)
-	require.NoError(t, err)
-
-	// Set birthday timestamp without setting birthday block.
-	birthdayTime := time.Now().UTC().Add(-30 * 24 * time.Hour)
-	updateParams := db.UpdateWalletParams{
-		WalletID: created.ID,
-		Birthday: &birthdayTime,
-	}
-	err = store.UpdateWallet(t.Context(), updateParams)
-	require.NoError(t, err)
-
-	retrieved, err := store.GetWallet(t.Context(), created.Name)
-	require.NoError(t, err)
-	require.Equal(t, birthdayTime.Unix(), retrieved.Birthday.Unix())
-
-	// Assert fields that were not updated remain unchanged.
-	require.Equal(t, created.ID, retrieved.ID)
-	require.Equal(t, created.Name, retrieved.Name)
-	require.Equal(t, created.IsImported, retrieved.IsImported)
-	require.Equal(t, created.ManagerVersion, retrieved.ManagerVersion)
-	require.Equal(t, created.IsWatchOnly, retrieved.IsWatchOnly)
-	require.Nil(t, retrieved.BirthdayBlock)
-	require.Nil(t, retrieved.SyncedTo)
-}
-
-// TestUpdateWalletNotFound verifies that updating a non-existent wallet fails.
-func TestUpdateWalletNotFound(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-	before := store.StatsSnapshot()
-
-	updateParams := db.UpdateWalletParams{
-		WalletID: 99999, // Non-existent ID.
-	}
-
-	err := store.UpdateWallet(t.Context(), updateParams)
-	require.Error(t, err)
-	require.ErrorIs(t, err, db.ErrWalletNotFound)
-
-	after := store.StatsSnapshot()
-	require.Equal(t, before, after)
-}
-
-// TestGetEncryptedHDSeed verifies retrieving the encrypted HD seed.
-func TestGetEncryptedHDSeed(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	params := CreateWalletParamsFixture("seed-wallet")
-	expectedSeed := params.EncryptedMasterPrivKey
-
-	created, err := store.CreateWallet(t.Context(), params)
-	require.NoError(t, err)
-
-	seed, err := store.GetEncryptedHDSeed(t.Context(), created.ID)
-	require.NoError(t, err)
-	require.Equal(t, expectedSeed, seed)
-}
-
-// TestGetWalletSecrets verifies retrieving all stored wallet secret material.
-func TestGetWalletSecrets(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	params := CreateWalletParamsFixture("wallet-secrets-read")
-	created, err := store.CreateWallet(t.Context(), params)
-	require.NoError(t, err)
-
-	secrets, err := store.GetWalletSecrets(t.Context(), created.ID)
-	require.NoError(t, err)
-	require.NotNil(t, secrets)
-	require.Equal(t, params.MasterKeyPrivParams, secrets.MasterPrivParams)
-	require.Equal(
-		t, params.EncryptedCryptoPrivKey, secrets.EncryptedCryptoPrivKey,
-	)
-	require.Equal(
-		t, params.EncryptedCryptoScriptKey,
-		secrets.EncryptedCryptoScriptKey,
-	)
-	require.Equal(
-		t, params.EncryptedMasterPrivKey, secrets.EncryptedMasterHdPrivKey,
-	)
-}
-
-// TestGetWalletSecretsWatchOnly verifies watch-only wallets return their
-// stored secret row without promoting empty fields to not-found.
-func TestGetWalletSecretsWatchOnly(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	params := CreateWatchOnlyWalletParams("watch-only-wallet-secrets-read")
-	created, err := store.CreateWallet(t.Context(), params)
-	require.NoError(t, err)
-
-	secrets, err := store.GetWalletSecrets(t.Context(), created.ID)
-	require.NoError(t, err)
-	require.NotNil(t, secrets)
-	require.Empty(t, secrets.MasterPrivParams)
-	require.Empty(t, secrets.EncryptedCryptoPrivKey)
-	require.Empty(t, secrets.EncryptedMasterHdPrivKey)
-	require.Empty(t, secrets.EncryptedCryptoScriptKey)
-}
-
-// TestGetWalletSecretsNotFound verifies missing wallets map to
-// ErrWalletNotFound.
-func TestGetWalletSecretsNotFound(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-	before := store.StatsSnapshot()
-
-	secrets, err := store.GetWalletSecrets(t.Context(), 99999)
-	require.Nil(t, secrets)
-	require.ErrorIs(t, err, db.ErrWalletNotFound)
-
-	after := store.StatsSnapshot()
-	require.Equal(t, before, after)
-}
-
-// TestGetEncryptedHDSeedWatchOnly verifies that watch-only wallets
-// have no encrypted seed.
-func TestGetEncryptedHDSeedWatchOnly(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	params := CreateWatchOnlyWalletParams("watch-only-seed")
-	created, err := store.CreateWallet(t.Context(), params)
-	require.NoError(t, err)
-
-	seed, err := store.GetEncryptedHDSeed(t.Context(), created.ID)
-	require.Nil(t, seed, "watch-only wallets should not have HD seed")
-	require.ErrorIs(t, err, db.ErrSecretNotFound)
-}
-
-// TestWatchOnlyWalletRejectsWalletSecrets verifies that watch-only
-// wallets allow script-encryption material while still rejecting
-// private wallet secrets.
-func TestWatchOnlyWalletRejectsWalletSecrets(t *testing.T) {
-	t.Parallel()
-
-	t.Run("create with no private secrets succeeds", func(t *testing.T) {
-		t.Parallel()
-
-		store := NewTestStore(t)
-
-		params := CreateWatchOnlyWalletParams("watch-only-create-ok")
-		info, err := store.CreateWallet(t.Context(), params)
-		require.NoError(t, err)
-		require.NotNil(t, info)
-		require.True(t, info.IsWatchOnly)
-	})
-
-	t.Run(
-		"create with empty-but-non-nil private secrets succeeds",
-		func(t *testing.T) {
-			t.Parallel()
-
-			store := NewTestStore(t)
-
-			params := CreateWatchOnlyWalletParams("watch-only-create-empty")
-			params.MasterKeyPrivParams = []byte{}
-			params.EncryptedCryptoPrivKey = []byte{}
-			params.EncryptedMasterPrivKey = []byte{}
-
-			info, err := store.CreateWallet(t.Context(), params)
-			require.NoError(t, err)
-			require.NotNil(t, info)
-			require.True(t, info.IsWatchOnly)
-
-			seed, err := store.GetEncryptedHDSeed(t.Context(), info.ID)
-			require.Nil(t, seed)
-			require.ErrorIs(t, err, db.ErrSecretNotFound)
-		},
-	)
-
-	t.Run("create with script key only succeeds", func(t *testing.T) {
-		t.Parallel()
-
-		store := NewTestStore(t)
-
-		params := CreateWatchOnlyWalletParams("watch-only-create-script")
-		params.EncryptedCryptoScriptKey = RandomBytes(32)
-
-		info, err := store.CreateWallet(t.Context(), params)
-		require.NoError(t, err)
-		require.NotNil(t, info)
-		require.True(t, info.IsWatchOnly)
-
-		seed, err := store.GetEncryptedHDSeed(t.Context(), info.ID)
-		require.Nil(t, seed)
-		require.ErrorIs(t, err, db.ErrSecretNotFound)
-	})
-
-	t.Run("create with private secrets is rejected", func(t *testing.T) {
-		t.Parallel()
-
-		store := NewTestStore(t)
-
-		params := CreateWatchOnlyWalletParams("watch-only-create-reject")
-		params.MasterKeyPrivParams = RandomBytes(16)
-
-		_, err := store.CreateWallet(t.Context(), params)
-		require.Error(t, err)
-		require.ErrorIs(t, err, db.ErrWatchOnlyViolation)
-	})
-
-	t.Run("update with script key only succeeds", func(t *testing.T) {
-		t.Parallel()
-
-		store := NewTestStore(t)
-
-		created, err := store.CreateWallet(
-			t.Context(),
-			CreateWatchOnlyWalletParams("watch-only-update-script"),
-		)
-		require.NoError(t, err)
-
-		err = store.UpdateWalletSecrets(
-			t.Context(), db.UpdateWalletSecretsParams{
-				WalletID:                 created.ID,
-				EncryptedCryptoScriptKey: RandomBytes(32),
-			},
-		)
-		require.NoError(t, err)
-
-		seed, err := store.GetEncryptedHDSeed(t.Context(), created.ID)
-		require.Nil(t, seed)
-		require.ErrorIs(t, err, db.ErrSecretNotFound)
-	})
-
-	t.Run(
-		"update with empty-but-non-nil private secrets succeeds",
-		func(t *testing.T) {
-			t.Parallel()
-
-			store := NewTestStore(t)
-
-			created, err := store.CreateWallet(
-				t.Context(),
-				CreateWatchOnlyWalletParams("watch-only-update-empty"),
-			)
-			require.NoError(t, err)
-
-			err = store.UpdateWalletSecrets(
-				t.Context(), db.UpdateWalletSecretsParams{
-					WalletID:                 created.ID,
-					MasterPrivParams:         []byte{},
-					EncryptedCryptoPrivKey:   []byte{},
-					EncryptedMasterHdPrivKey: []byte{},
-				},
-			)
-			require.NoError(t, err)
-
-			seed, err := store.GetEncryptedHDSeed(t.Context(), created.ID)
-			require.Nil(t, seed)
-			require.ErrorIs(t, err, db.ErrSecretNotFound)
-		},
-	)
-
-	t.Run("update with private secrets is rejected", func(t *testing.T) {
-		t.Parallel()
-
-		store := NewTestStore(t)
-
-		created, err := store.CreateWallet(
-			t.Context(),
-			CreateWatchOnlyWalletParams("watch-only-update-reject"),
-		)
-		require.NoError(t, err)
-
-		err = store.UpdateWalletSecrets(
-			t.Context(), db.UpdateWalletSecretsParams{
-				WalletID:                 created.ID,
-				MasterPrivParams:         RandomBytes(16),
-				EncryptedCryptoPrivKey:   RandomBytes(32),
-				EncryptedMasterHdPrivKey: RandomBytes(32),
-			},
-		)
-		require.Error(t, err)
-		require.ErrorIs(t, err, db.ErrWatchOnlyViolation)
-	})
-}
-
-// TestWatchOnlyWalletSecretTriggers verifies that wallet_secrets rejects
-// watch-only parent wallets while still allowing inserts and updates for
-// non-watch-only parents.
-func TestWatchOnlyWalletSecretTriggers(t *testing.T) {
-	t.Parallel()
-
-	t.Run("watch-only insert is rejected", func(t *testing.T) {
-		t.Parallel()
-
-		store := NewTestStore(t)
-
-		walletInfo, err := store.CreateWallet(
-			t.Context(),
-			CreateWatchOnlyWalletParams("watch-only-wallet-secret-insert"),
-		)
-		require.NoError(t, err)
-
-		err = insertWalletSecretRaw(
-			t, store.DB(), walletInfo.ID, RandomBytes(16), RandomBytes(32),
-			RandomBytes(32), RandomBytes(32),
-		)
-		require.Error(t, err)
-		requireDriverConstraintError(t, err)
-	})
-
-	t.Run("watch-only script-only insert succeeds", func(t *testing.T) {
-		t.Parallel()
-
-		store := NewTestStore(t)
-
-		walletInfo, err := store.CreateWallet(
-			t.Context(),
-			CreateWatchOnlyWalletParams("watch-only-wallet-secret-script-only"),
-		)
-		require.NoError(t, err)
-
-		err = deleteWalletSecretRaw(t, store.DB(), walletInfo.ID)
-		require.NoError(t, err)
-
-		err = insertWalletSecretRaw(
-			t, store.DB(), walletInfo.ID, nil, nil, RandomBytes(32), nil,
-		)
-		require.NoError(t, err)
-	})
-
-	t.Run(
-		"watch-only empty-but-non-nil insert is rejected",
-		func(t *testing.T) {
-			t.Parallel()
-
-			store := NewTestStore(t)
-
-			walletInfo, err := store.CreateWallet(
-				t.Context(),
-				CreateWatchOnlyWalletParams(
-					"watch-only-wallet-secret-empty-insert",
-				),
-			)
-			require.NoError(t, err)
-
-			err = insertWalletSecretRaw(
-				t, store.DB(), walletInfo.ID, []byte{}, nil, RandomBytes(32),
-				nil,
-			)
-			require.Error(t, err)
-			requireDriverConstraintError(t, err)
-		},
-	)
-
-	t.Run("watch-only update is rejected", func(t *testing.T) {
-		t.Parallel()
-
-		store := NewTestStore(t)
-
-		params := CreateWatchOnlyWalletParams("watch-only-wallet-secret-update")
-		params.EncryptedCryptoScriptKey = RandomBytes(32)
-
-		walletInfo, err := store.CreateWallet(t.Context(), params)
-		require.NoError(t, err)
-
-		seed, err := store.GetEncryptedHDSeed(t.Context(), walletInfo.ID)
-		require.Nil(t, seed)
-		require.ErrorIs(t, err, db.ErrSecretNotFound)
-
-		err = updateWalletSecretRaw(
-			t, store.DB(), walletInfo.ID, nil, nil, RandomBytes(32), nil,
-		)
-		require.NoError(t, err)
-
-		err = updateWalletSecretRaw(
-			t, store.DB(), walletInfo.ID, RandomBytes(16), RandomBytes(32),
-			RandomBytes(32), RandomBytes(32),
-		)
-		require.Error(t, err)
-		requireDriverConstraintError(t, err)
-
-		err = updateWalletSecretRaw(
-			t, store.DB(), walletInfo.ID, []byte{}, nil, RandomBytes(32),
-			nil,
-		)
-		require.Error(t, err)
-		requireDriverConstraintError(t, err)
-	})
-
-	t.Run("non-watch-only insert and update succeed", func(t *testing.T) {
-		t.Parallel()
-
-		store := NewTestStore(t)
-
-		walletID := newWallet(t, store, "spendable-wallet-secret-trigger")
-
-		err := deleteWalletSecretRaw(t, store.DB(), walletID)
-		require.NoError(t, err)
-
-		insertedSeed := RandomBytes(32)
-		err = insertWalletSecretRaw(
-			t, store.DB(), walletID, RandomBytes(16), RandomBytes(32),
-			RandomBytes(32), insertedSeed,
-		)
-		require.NoError(t, err)
-
-		seed, err := store.GetEncryptedHDSeed(t.Context(), walletID)
-		require.NoError(t, err)
-		require.Equal(t, insertedSeed, seed)
-
-		updatedSeed := RandomBytes(32)
-		err = updateWalletSecretRaw(
-			t, store.DB(), walletID, RandomBytes(16), RandomBytes(32),
-			RandomBytes(32), updatedSeed,
-		)
-		require.NoError(t, err)
-
-		seed, err = store.GetEncryptedHDSeed(t.Context(), walletID)
-		require.NoError(t, err)
-		require.Equal(t, updatedSeed, seed)
-	})
-}
-
-// TestWalletWatchOnlyImmutable verifies that raw wallet updates cannot change
-// the watch-only flag after wallet creation.
-func TestWalletWatchOnlyImmutable(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name             string
-		walletName       string
-		params           func(string) db.CreateWalletParams
-		updatedWatchOnly bool
-	}{
-		{
-			name:             "spendable wallet cannot become watch-only",
-			walletName:       "wallet-watch-only-immutable-spendable",
-			params:           CreateWalletParamsFixture,
-			updatedWatchOnly: true,
-		},
-		{
-			name:             "watch-only wallet cannot become spendable",
-			walletName:       "wallet-watch-only-immutable-watch-only",
-			params:           CreateWatchOnlyWalletParams,
-			updatedWatchOnly: false,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-
-			store := NewTestStore(t)
-
-			created, err := store.CreateWallet(
-				t.Context(), tc.params(tc.walletName),
-			)
-			require.NoError(t, err)
-
-			err = updateWalletWatchOnlyRaw(
-				t, store.DB(), created.ID, tc.updatedWatchOnly,
-			)
-			require.Error(t, err)
-			requireDriverConstraintError(t, err)
-
-			walletInfo, err := store.GetWallet(t.Context(), tc.walletName)
-			require.NoError(t, err)
-			require.NotNil(t, walletInfo)
-			require.Equal(t, created.IsWatchOnly, walletInfo.IsWatchOnly)
-		})
-	}
-}
-
-// TestUpdateWalletSecrets checks that updating the wallet secrets works
-// correctly.
-func TestUpdateWalletSecrets(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	params := CreateWalletParamsFixture("secrets-wallet")
-	created, err := store.CreateWallet(t.Context(), params)
-	require.NoError(t, err)
-
-	newSecrets := db.UpdateWalletSecretsParams{
-		WalletID:                 created.ID,
-		MasterPrivParams:         RandomBytes(16),
-		EncryptedCryptoPrivKey:   RandomBytes(32),
-		EncryptedCryptoScriptKey: RandomBytes(32),
-		EncryptedMasterHdPrivKey: RandomBytes(32),
-	}
-
-	err = store.UpdateWalletSecrets(t.Context(), newSecrets)
-	require.NoError(t, err)
-
-	seed, err := store.GetEncryptedHDSeed(t.Context(), created.ID)
-	require.NoError(t, err)
-	require.Equal(t, newSecrets.EncryptedMasterHdPrivKey, seed)
-}
-
-// TestUpdateWalletAutoBlockInsertion verifies that UpdateWallet automatically
-// inserts blocks when updating SyncedTo or BirthdayBlock.
-func TestUpdateWalletAutoBlockInsertion(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	params := CreateWalletParamsFixture("auto-block-wallet")
-	created, err := store.CreateWallet(t.Context(), params)
-	require.NoError(t, err)
-
-	// Create a block WITHOUT pre-inserting it into the blocks table.
-	block := db.Block{
-		Height:    uint32(100),
-		Hash:      RandomHash(),
-		Timestamp: time.Now().UTC(),
-	}
-
-	// Update wallet with SyncedTo - should automatically insert the block.
-	updateParams := db.UpdateWalletParams{
-		WalletID: created.ID,
-		SyncedTo: &block,
-	}
-	err = store.UpdateWallet(t.Context(), updateParams)
-	require.NoError(t, err)
-
-	// Verify the wallet was updated.
-	retrieved, err := store.GetWallet(t.Context(), created.Name)
-	require.NoError(t, err)
-	require.NotNil(t, retrieved.SyncedTo)
-	require.Equal(t, block.Height, retrieved.SyncedTo.Height)
-	require.Equal(t, block.Hash, retrieved.SyncedTo.Hash)
-
-	// Update again with the same block - should be idempotent.
-	err = store.UpdateWallet(t.Context(), updateParams)
-	require.NoError(t, err, "updating with same block should be idempotent")
-
-	// Create another block for BirthdayBlock.
-	birthdayBlock := db.Block{
-		Height:    uint32(50),
-		Hash:      RandomHash(),
-		Timestamp: time.Now().UTC().Add(-time.Hour),
-	}
-
-	// Update wallet with BirthdayBlock - should automatically insert it.
-	updateParams = db.UpdateWalletParams{
-		WalletID:      created.ID,
-		BirthdayBlock: &birthdayBlock,
-	}
-	err = store.UpdateWallet(t.Context(), updateParams)
-	require.NoError(t, err)
-
-	// Verify both blocks are set.
-	retrieved, err = store.GetWallet(t.Context(), created.Name)
-	require.NoError(t, err)
-	require.NotNil(t, retrieved.SyncedTo)
-	require.NotNil(t, retrieved.BirthdayBlock)
-	require.Equal(t, block.Height, retrieved.SyncedTo.Height)
-	require.Equal(t, birthdayBlock.Height, retrieved.BirthdayBlock.Height)
-}
